+++ CHANGELOG.md
@@ -23,7 +23,8 @@
 - Set sideEffects to false.
 - The type of "this" was made explicit.
 - weekStart in weeksInYear.
-- Set DateTime default year to 1970
+- Set DateTime default year to 1970.
+- Add day to the normalizeTime result.
 
 ### Removed